+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 ## v0.2.7
 
 * Add setting to control which types of objects from the current projects are included when offering completions.
+* _Experimental:_ Add completions for qualified and unqualified item metadata expressions (`%(XXX.YYY)` and `%(YYY)`).
 
 ## v0.2.6
 
 * Bug-fix: attribute completions should be available on elements that don't currently have any attributes.
 * Bug-fix: go-to-definition should also work for regular-style project imports (not just SDK-style imports).
-* _Experimental:_ Add completions for MSBuild property  and item expressions (`$()` and `@()`).
+* _Experimental:_ Add completions for MSBuild property and item expressions (`$(XXX)` and `@(XXX)`).
 
 ## v0.2.4